Menu add separated as independent component, correct output for Lead detail props
