+++ src/aliance/List/index.jsx
@@ -2,7 +2,7 @@ import React, { Fragment } from "react";
 import { connect } from "react-redux";
 import { Link } from "react-router-dom";
 
-import MenuAdd from "../Menu/Add";
+import MenuAdd from "../../Menu/Add";
 import Balance from "../Balance";
 
 import { Button } from "material-ui";
